Corrections

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,16 +1,17 @@
 #include "search_algos.h"
 #include <stdio.h>
 #include <stdlib.h>
-/*
- * linear_search - a function to search for a value in an array
- * array: pointer to first element of array
- * size: the number of elements in array
- * value: value to search for
- * Return: first index where value is located
+/**
+ * linear_search - linear search algorithm for arrays.
+ * @array: pointer to first element of array.
+ * @size: the number of elements in array.
+ * @value: value to search for.
+ * Return: first index where value is located.
  */
 int linear_search(int *array, size_t size, int value)
 {
-	long unsigned int i;
+	size_t i;
+
 	if (!array || size <= 0)
 		return (-1);
 	for (i = 0; i < size; i++)
@@ -19,5 +20,5 @@ int linear_search(int *array, size_t size, int value)
 		if (array[i] == value)
 			return (i);
 	}
-	return (-1);			
+	return (-1);
 }
